#11 Added a defined time frame for polling servers over GSM
The design of the GSM client object is such that the polling
of data continuously after a request has been made, is necessary.

To combat this without the need for having the initial connection
code in `setup()`,  a while loop calls these relevant functions for
the set period of time as maximum after exhausting.