AzureIotHub_MKRGSM library / singleton class
The IotHubClient class in AzureIotHub_MKRGSM is designed to provide
an easy interface for MQTT communications over SSL using the GSM
module for the Arduino MKR 1400.

The user need only provide the instance of the class with the address
for the mqtt broker, and the device ID which is coupled with the device
id in the users' IotHub, and start publishing messages.

*Refer to the docstrings in the class for individual documentation
for each method.*