Bug fix - GPS connectivity issues SOLVED
Connectivity would suffer with the GPS shield in conjunction with the
GSM connectivity with Azure over MQTT.

This was ultimately caused by lack of iterations over the `GPS.available`
method, where the `IotHubClient::Update()` method was called before or after,
on the same function call to `loop` which did not leave enough time for
these two polling methods respectively.

The bug was solved by allowing `GPS.available()` an allowance of
cycles to perform polling, whereafter the coordinates are sent to
IoT Hub.

Fixes #19