Improvement to bug #19 and GPS connectivity
The problem with establishing a connection to the GPS is evident as
providing the MCU with a set amount of cycles to poll the GPS for a lock
is insufficient at best.

This was remedied by having a set time of connectivity attempts, just as
with GSM and MQTT.

This parameter is changeable in defines.h.