Password class
The Password class is designed to encapsulate secure representation
of a password between the front end of the app and the back end with
database integration.

The plain text password is never stored but instantly hashed using
SHA3-512 with a pseudorandom salt.

The salt and hash are stored in the Password object.

Pythonic dunder methods __eq__ and __str__ make it very easy both
to see the hash of the password, and to verify a plain string against
the hashed password.