User class
the User class represents a user, with the typical attributes associated
with a user in applications.

The User class hold regular properties such as name and username, email.

The instance will assign a UUID upon CTOR automatically.

The instance has an instance of the <Password> class, to securely
encapsulate the password for the user in a hashed format.