#24 Refactor of the SqlCommand class
The SqlCommand class has been vastly refactored with many improvements.

The SqlCommand now offers a complete pythonic approach to SQL
queries, using dicts and tuples for parameters such as columns and
values. OR, AND, WHERE, INNER/OUTER/LEFT/RIGHT/joins are supported.

The __repr__ presents the instance as a complete and concatenated
SQL command using the parameters with variables provided.