New class: SqlCondition
The SqlCondition class, when instantiated, constructs a syntactically
correct SQL WHERE, SET, or ON statement separated by " = ", for use
in the SqlCommand object when these types of conditions are necessary.

The move from the stdlib 'dict' object was necessary as OR statements
are impossible where 'dict' cannot hold multiple entries of keys.

This class operates in a syntactically similar way as 'dict' but
with the sole purpose of string concatenation in SQL fashion.