The SqlCondition class
The SqlCondition class is a key-value-like object to store
SQL conditions for SET, ON and WHERE statements (or anywhere else where
'x = y' syntax is desired), with support for 'AND' and 'OR' to combine
more complex statements in one object.

The SqlCondition replaces the previously used 'dict' due to the
nature of hashmaps where keys are unique.

This obviously causes issues when 'OR' statments are used since the
key name must be able to occur more than once.

The SqlCondition is designed to be used with the SqlCommand class
to supplement its conditional statements for an easy OOP and pythonic
approach to declare these statements.